Move composercontent to components/dialog
