+++ gsa/src/web/components/dialog/composercontent.js
