+++ gsa/src/web/components/reportcontentcomposer/triggeralertdialog.js
@@ -29,6 +29,7 @@ import {isString} from 'gmp/utils/identity';
 import PropTypes from 'web/utils/proptypes';
 import {renderSelectItems} from 'web/utils/render';
 
+import ComposerContent from 'web/components/dialog/composercontent';
 import SaveDialog from 'web/components/dialog/savedialog';
 
 // import CheckBox from 'web/components/form/checkbox';
@@ -39,8 +40,6 @@ import NewIcon from 'web/components/icon/newicon';
 
 import Divider from 'web/components/layout/divider';
 import Layout from 'web/components/layout/layout';
-
-import ComposerContent from './composercontent';
 
 // const StyledDiv = styled.div`
 //   text-align: end;
